Split chart/tests.rs to stay under 1000-line limit
chart/tests.rs was at 1115 lines, the only file in the project
exceeding the 1000-line file size rule. Moves snapshot/rendering
tests into a separate file.
